requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
pyodbc==4.0.39
python-engineio==4.8.0
python-socketio==5.10.0
Werkzeug==2.3.7 